+++ lib/td_dd/grants/search.ex
@@ -10,8 +10,6 @@ defmodule TdDd.Grants.Search do
 
   require Logger
 
-  @behaviour Bodyguard.Policy
-
   @index :grants
   @default_sort ["_id"]
   @aggs %{
@@ -22,8 +20,6 @@ defmodule TdDd.Grants.Search do
       terms: %{field: "data_structure_version.system.external_id.raw", size: 50}
     }
   }
-
-  def authorize(:manage, %{role: role}, _any), do: role == "admin"
 
   def get_filter_values(claims, params, user_id) do
     params = put_filter(params, "user_id", user_id)