+++ test/td_dd_web/controllers/grant_request_filter_controller_test.exs
@@ -1,0 +1,104 @@
+defmodule TdDdWeb.GrantRequestFilterControllerTest do
+  use TdDdWeb.ConnCase
+
+  import Mox
+
+  @aggregations %{
+    "foo" => %{
+      "buckets" => [%{"key" => "bar"}, %{"key" => "baz"}]
+    }
+  }
+
+  setup_all do
+    start_supervised(TdDd.Search.Cluster)
+    :ok
+  end
+
+  setup :verify_on_exit!
+
+  describe "POST /api/grant_requests_filters/search" do
+    @tag authentication: [role: "admin"]
+    test "includes a match_all filters for admin", %{conn: conn} do
+      ElasticsearchMock
+      |> expect(:request, fn
+        _, :post, "/grant_requests/_search", %{query: query, size: 0, aggs: _}, _ ->
+          assert query == %{bool: %{filter: %{match_all: %{}}}}
+
+          SearchHelpers.aggs_response(@aggregations)
+      end)
+
+      assert %{"data" => data} =
+               conn
+               |> post(Routes.grant_request_filter_path(conn, :search, %{}))
+               |> json_response(:ok)
+
+      assert %{"foo" => %{"values" => ["bar", "baz"]}} = data
+    end
+
+    @tag authentication: [user_name: "non_admin_user", permissions: ["approve_grant_request"]]
+    test "filters by permissions with non-admin user", %{conn: conn} do
+      ElasticsearchMock
+      |> expect(:request, fn
+        _, :post, "/grant_requests/_search", %{query: query, size: 0, aggs: _}, _ ->
+          assert %{
+                   bool: %{
+                     filter: %{
+                       bool: %{
+                         should: [
+                           %{term: %{"data_structure_version.domain_ids" => _}}
+                         ]
+                       }
+                     }
+                   }
+                 } = query
+
+          SearchHelpers.aggs_response()
+      end)
+
+      assert %{"data" => %{}} =
+               conn
+               |> post(Routes.grant_request_filter_path(conn, :search, %{}))
+               |> json_response(:ok)
+    end
+  end
+
+  @tag authentication: [user_name: "non_admin_user", permissions: ["approve_grant_request"]]
+  test "includes filerts from request parameters", %{conn: conn} do
+    ElasticsearchMock
+    |> expect(:request, fn
+      _, :post, "/grant_requests/_search", %{query: query, size: 0}, _ ->
+        assert %{
+                 bool: %{
+                   filter: [
+                     %{term: %{"foo" => "bar"}},
+                     _permission_filter
+                   ]
+                 }
+               } = query
+
+        SearchHelpers.aggs_response()
+    end)
+
+    params = %{"filters" => %{"foo" => ["bar"]}}
+
+    assert %{"data" => %{}} =
+             conn
+             |> post(Routes.grant_request_filter_path(conn, :search, params))
+             |> json_response(:ok)
+  end
+
+  @tag authentication: [user_name: "non_admin"]
+  test "not include filerts if user dont have permissions", %{conn: conn} do
+    ElasticsearchMock
+    |> expect(:request, fn
+      _, :post, "/grant_requests/_search", %{query: query, size: 0, aggs: _}, _ ->
+        assert %{bool: %{filter: %{match_none: %{}}}} = query
+        SearchHelpers.aggs_response()
+    end)
+
+    assert %{"data" => %{}} =
+             conn
+             |> post(Routes.grant_request_filter_path(conn, :search, %{}))
+             |> json_response(:ok)
+  end
+end