+++ test/td_dq_web/controllers/implementation_upload_controller_test.exs
@@ -583,6 +583,76 @@ defmodule TdDdWeb.ImplementationUploadControllerTest do
     end
 
     @tag authentication: [role: "admin"]
+    test "upload implementations in native language", %{conn: conn} do
+      attrs = %{
+        implementations: %Plug.Upload{
+          filename: "implementations_translations.csv",
+          path: "test/fixtures/implementations/implementations_translations.csv"
+        },
+        lang: "es"
+      }
+
+      template_content = [
+        %{
+          "fields" => [
+            %{
+              "cardinality" => "1",
+              "label" => "label_i18n",
+              "name" => "i18n",
+              "type" => "string",
+              "values" => %{"fixed" => ["one", "two", "three"]}
+            }
+          ],
+          "name" => "group_name0"
+        }
+      ]
+
+      CacheHelpers.insert_template(
+        name: "i18n_template",
+        scope: "ri",
+        content: template_content
+      )
+
+      CacheHelpers.put_i18n_message("es", %{
+        message_id: "fields.label_i18n.one",
+        definition: "uno"
+      })
+
+      CacheHelpers.put_i18n_message("es", %{
+        message_id: "fields.label_i18n.two",
+        definition: "dos"
+      })
+
+      CacheHelpers.put_i18n_message("es", %{
+        message_id: "fields.label_i18n.three",
+        definition: "tres"
+      })
+
+      CacheHelpers.put_i18n_message("es", %{
+        message_id: "ruleImplementations.props.result_type.deviation",
+        definition: "Desviación"
+      })
+
+      CacheHelpers.put_i18n_message("es", %{
+        message_id: "ruleImplementations.props.result_type.errors_number",
+        definition: "Número"
+      })
+
+      CacheHelpers.put_i18n_message("es", %{
+        message_id: "ruleImplementations.props.result_type.percentage",
+        definition: "Porcentaje"
+      })
+
+      assert %{"data" => data} =
+               conn
+               |> post(Routes.implementation_upload_path(conn, :create), attrs)
+               |> json_response(:ok)
+
+      assert %{"ids" => ids, "errors" => []} = data
+      assert length(ids) == 3
+    end
+
+    @tag authentication: [role: "admin"]
     test "uploads implementations without rules", %{conn: conn} do
       CacheHelpers.insert_domain(external_id: "some_domain_id")